+++ tests/suites/temporal.cpp
@@ -11,6 +11,8 @@
 namespace units = ::JadeMatrix::units;
 
 
+// Tests use `long double` values as these are only checking for numerically
+// correct conversions (not type semantics preservation)
 
 TEST_CASE( "stream format seconds" )
 {
@@ -26,7 +28,6 @@ TEST_CASE( "std::chrono::seconds to JadeMatrix::units::seconds" )
     units::seconds< rep > seconds{ std_seconds };
     REQUIRE( static_cast< rep >( seconds ) == 7 );
 }
-
 TEST_CASE( "std::chrono::minutes to JadeMatrix::units::seconds" )
 {
     std::chrono::minutes std_minutes{ 7 };
@@ -34,7 +35,6 @@ TEST_CASE( "std::chrono::minutes to JadeMatrix::units::seconds" )
     units::seconds< rep > seconds{ std_minutes };
     REQUIRE( static_cast< rep >( seconds ) == 420 );
 }
-
 TEST_CASE( "JadeMatrix::units::seconds to std::chrono::seconds" )
 {
     units::seconds< long long > seconds{ 7 };
@@ -42,46 +42,40 @@ TEST_CASE( "JadeMatrix::units::seconds to std::chrono::seconds" )
     REQUIRE( std_seconds.count() == 7 );
 }
 
-
 TEST_CASE( "seconds to minutes" )
 {
-    units::seconds< double > seconds{ 7 };
-    units::minutes< double > minutes{ seconds };
-    REQUIRE( static_cast< double >( minutes ) == ( 7.0 / 60.0 ) );
+    units::seconds< long double > seconds{ 7 };
+    units::minutes< long double > minutes{ seconds };
+    REQUIRE( static_cast< long double >( minutes ) == ( 7.0L / 60.0L ) );
 }
-
 TEST_CASE( "seconds to hours" )
 {
-    units::seconds< double > seconds{ 7 };
-    units::hours< double > hours{ seconds };
-    REQUIRE( static_cast< double >( hours ) == ( 7.0 / 3600.0 ) );
+    units::seconds< long double > seconds{ 7 };
+    units::hours< long double > hours{ seconds };
+    REQUIRE( static_cast< long double >( hours ) == ( 7.0L / 3600.0L ) );
 }
-
 TEST_CASE( "minutes to hours" )
 {
-    units::minutes< double > minutes{ 7 };
-    units::hours< double > hours{ minutes };
-    REQUIRE( static_cast< double >( hours ) == ( 7.0 / 60.0 ) );
+    units::minutes< long double > minutes{ 7 };
+    units::hours< long double > hours{ minutes };
+    REQUIRE( static_cast< long double >( hours ) == ( 7.0L / 60.0L ) );
 }
-
 
 TEST_CASE( "minutes to seconds" )
 {
-    units::minutes< double > minutes{ 7 };
-    units::seconds< double > seconds{ minutes };
-    REQUIRE( static_cast< double >( seconds ) == 420.0 );
+    units::minutes< long double > minutes{ 7 };
+    units::seconds< long double > seconds{ minutes };
+    REQUIRE( static_cast< long double >( seconds ) == 420.0L );
 }
-
 TEST_CASE( "hours to seconds" )
 {
-    units::hours< double > hours{ 7 };
-    units::seconds< double > seconds{ hours };
-    REQUIRE( static_cast< double >( seconds ) == 25200.0 );
+    units::hours< long double > hours{ 7 };
+    units::seconds< long double > seconds{ hours };
+    REQUIRE( static_cast< long double >( seconds ) == 25200.0L );
 }
-
 TEST_CASE( "hours to minutes" )
 {
-    units::hours< double > hours{ 7 };
-    units::minutes< double > minutes{ hours };
-    REQUIRE( static_cast< double >( minutes ) == 420.0 );
+    units::hours< long double > hours{ 7 };
+    units::minutes< long double > minutes{ hours };
+    REQUIRE( static_cast< long double >( minutes ) == 420.0L );
 }
